pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marco-agent"
version = "0.1.0"
description = "Marco, an Azure-native AI Chief of Staff for Discord."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.10.11",
  "azure-cosmos>=4.7.0",
  "azure-functions>=1.21.3",
  "azure-identity>=1.18.0",
  "azure-keyvault-secrets>=4.9.0",
  "azure-storage-blob>=12.23.1",
  "discord.py>=2.5.2",
  "feedparser>=6.0.11",
  "openai>=1.70.0",
  "opencensus-ext-azure>=1.1.13",
  "pydantic>=2.11.1",
  "pypdf>=5.4.0",
  "python-docx>=1.1.2",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "ruff>=0.11.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
